+++ plugins/pong.go
@@ -1,20 +1,16 @@
 package plugins
 
 import (
-	"log"
-
 	dg "github.com/bwmarrin/discordgo"
 )
 
 // Pong returns the string "Pong" when a user types "--Ping"
-func (r *Record) Pong(s *dg.Session, msg *dg.MessageCreate) {
+func (r *Record) Pong(s *dg.Session, msg *dg.MessageCreate) string {
 	// Check the last time the user made this request
-	if tooSoon := r.checkLastAsk(s, msg); tooSoon {
-		return
+	alertUser, tooSoon := r.checkLastAsk(s, msg)
+	if tooSoon {
+		return alertUser
 	}
-	_, err := s.ChannelMessageSend(msg.ChannelID, "pong")
-	if err != nil {
-		log.Fatalf("session.ChannelMessageSend failed: %s", err)
-	}
-	log.Printf("%s fetched pong", msg.Author.Username)
+
+	return "pong"
 }